Sign 폴더 및 SignUpSetting, SignInSetting 추가
